Fixed error when creating a repository with an indirect dependency on Common_Foundation
There was an error with dependencies that didn't directly depend on Common_Foundation, but rather had a dependency that ultimately had a dependency on Common_Foundation that resulted in a key error when attempting to walk a prioritized list of all repositories (as the search for repos during setup will bail once all of its direct dependencies are found).

This change introduces a condition around that scenario where it will not attempt to lookup dependencies of dependencies when recurse is not specified. This also addresses a problem when attempting to compare the "fingerprints" of dependencies (as the list of dependencies is different at setup time (only direct dependencies) and activation time (all dependencies)).