🐛 [-bug] Fixed regression
Fixed regression associated with 7bec2c99d3f22403cd50d9fe6e6bb04802f62bf3.

Repos that have a direct dependency on Common_Foundation will have a valid link to that repo, those that don't have a direct dependency will not. The fix is to use that information if we have it, but revert to prior behavior if we don't (which should be ok in all scenarios, as indirect dependencies will not be impacted by the issue that was addressed by the previous change).