👷 [+ci] Removed dependencies (3 of N)
Removed the validation of dependency repositories so that the 'main_stable' tag can be updated to include 040452171964d4821adb5f45e1bc62ea377adc7c . Normally, this kind of hack wouldn't be necessary. But, the bug that the referenced commit addresses is encountered during the bootstrap process invoked before Common_Foundation has been updated. Therefore, the fix never makes it to the dependency repository and their workflow fails. That workflow failure means that this change never makes it to 'main_stable'.

Disable the dependencies, allow 'main_stable' to be updated to include the fix, and then restore the dependencies here.