🐛 [-bug] Require bash shell during activation
All activation scripts require bash at this time. Normally, linux handles this by respecting the shebang directive. However, this directive is ignored when the script is sourced directly (as the currently active shell is used in all cases). This presents a problem on MacOS, as the default shell is now zsh. Issue an error and bail.